move repetitive testing into test_utils
